API

API

Trace.enable — Function.

Trace.enable()

Enables logging for all subsequent tracing macros and adds a "trace" (5) logging level.

source

Trace.@alert — Macro.

@alert(args...)

Executes alert(args...) only when tracing is enabled (Trace.enable()).

source

Trace.@critical — Macro.

@critical(args...)

Executes critical(args...) only when tracing is enabled (Trace.enable()).

source

Trace.@debug — Macro.

@debug(args...)

Executes debug(args...) only when tracing is enabled (Trace.enable()).

source

Trace.@emergency — Macro.

@emergency(args...)

Executes emergency(args...) only when tracing is enabled (Trace.enable()).

source

Trace.@error — Macro.

@error(args...)

Executes error(args...) only when tracing is enabled (Trace.enable()).

source

Trace.@info — Macro.

@info(args...)

Executes info(args...) only when tracing is enabled (Trace.enable()).

source

Trace.@log — Macro.

@log(args...)

Executes log(args...) only when tracing is enabled (Trace.enable()).

source

Trace.@notice — Macro.

@notice(args...)

Executes notice(args...) only when tracing is enabled (Trace.enable()).

source

Trace.@trace — Macro.

@trace(logger, ex)

Monitors execution time and memory allocation for the expression and automatically logs those metrics to the logger.

source

Trace.@warn — Macro.

@warn(args...)

Executes warn(args...) only when tracing is enabled (Trace.enable()).

source

Trace.TraceRecord — Type.

TraceRecord <: Record

Stores extra metrics (e.g., execution time, memory allocated) about a function call for trace records.

source

Trace.trace — Function.

trace(f, logger, args...; kwargs...)

Wrapper function which measures the executions time and memory allocation of function f(args...; kwargs...) and logs the resulting TraceRecord.

source